fix: rename webhook_app dir to app dir
