+++ src/app/controllers/measurement/stop/measurement_stop_controller.ts
