+++ src/app/view_models/reply/reply_view_model.ts
